+++ lib/convert/controller/convert_controller.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+import '../../portfolio/model/coin_view_data.dart';
+
+class AllCoinsController extends ChangeNotifier {
+  late List<CoinViewData> coins;
+}
