+++ lib/convert/provider/conver_provider.dart
@@ -1,0 +1,6 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import '../controller/convert_controller.dart';
+
+final allCoinsControllerProvider =
+    ChangeNotifierProvider((ref) => AllCoinsController());
