+++ lib/detail/mapper/coin_price_mapper.dart
@@ -1,5 +1,5 @@
-import 'package:warren_task_one/shared/api/models/coin_prices/coin_price_response.dart';
-import 'package:warren_task_one/shared/api/models/coin_prices/get_coin_price_response.dart';
+import '../../shared/api/models/coin_prices/coin_price_response.dart';
+import '../../shared/api/models/coin_prices/get_coin_price_response.dart';
 
 extension CoinMapper on GetCoinPriceResponse {
   List<CoinValueResponse> toViewData() {
