+++ lib/detail/usecase/get_coin_prices_use_case.dart
@@ -1,8 +1,7 @@
-import 'package:warren_task_one/detail/mapper/coin_price_mapper.dart';
-
 import '../../shared/api/models/coin_prices/coin_price_response.dart';
 import '../../shared/repository/coin_prices_repository.dart';
 import '../controller/detail_controller.dart';
+import '../mapper/coin_price_mapper.dart';
 
 class GetCoinPricesUseCase {
   final CoinPricesRepository repository;
