+++ lib/detail/widgets/convert_button.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:warren_task_one/convert/view/convert_page.dart';
-import 'package:warren_task_one/core/app_routes.dart';
-
-import 'package:warren_task_one/portfolio/model/coin_view_data.dart';
+import '../../convert/view/convert_page.dart';
+import '../../portfolio/model/coin_view_data.dart';
 
 import '../../shared/api/models/coin_prices/coin_price_response.dart';
