+++ lib/detail/widgets/detail_description.dart
@@ -3,7 +3,7 @@ import 'package:decimal/intl.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
-import 'package:warren_task_one/shared/api/models/coin_prices/coin_price_response.dart';
+import '../../shared/api/models/coin_prices/coin_price_response.dart';
 
 import '../../portfolio/model/wallet_view_data.dart';
 import '../../shared/utils/util.dart';
