+++ lib/portfolio/mapper/coin_mapper.dart
@@ -1,5 +1,5 @@
-import '../model/coin_view_data.dart';
 import '../../shared/api/models/coin/get_all_coins_response.dart';
+import '../model/coin_view_data.dart';
 
 extension CoinMapper on GetAllCoinsResponse {
   List<CoinViewData> toViewData() {
