+++ lib/portfolio/provider/wallet_provider.dart
@@ -14,8 +14,6 @@ final pageIndexProvider = StateProvider<int>(
 
 final walletControllerProvider =
     ChangeNotifierProvider((ref) => WalletController());
-final allCoinsControllerProvider =
-    ChangeNotifierProvider((ref) => AllCoinsController());
 
 final getAllCoinUseCase = Provider((ref) {
   return GetAllCoinUseCase(repository: ref.read(coinRepositoryProvider));
