+++ lib/portfolio/widgets/portfolio_body.dart
@@ -1,7 +1,8 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:warren_task_one/portfolio/repository/wallet_repository.dart';
+import '../../convert/provider/conver_provider.dart';
+import '../repository/wallet_repository.dart';
 import 'coin_list.dart';
 import 'wallet_header.dart';
